js/src/xpconnect/tests/chrome/test_precisegc.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=533596
-->
<window title="Mozilla Bug 661927"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">

  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
      Components.utils.schedulePreciseGC(
        function() {
          ok(true, "callback executed");
          SimpleTest.finish();
        });
      SimpleTest.waitForExplicitFinish();
  ]]></script>
</window>
